pyproject.toml
[project]
name = "llm-engineering"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "groq>=0.31.0",
    "ipykernel>=6.30.1",
    "ipython>=9.5.0",
    "ipywidgets>=8.1.7",
    "langchain~=0.3.27",
    "langchain-google-genai>=2.1.10",
    "langchain-groq>=0.3.7",
    "langchain-huggingface>=0.3.1",
    "loguru>=0.7.3",
    "python-dotenv>=1.1.1",
    "sentence-transformers>=5.1.0",
    "streamlit>=1.49.1",
]
